// apps/web/src/lib/api/generated/types/updateConsultantDtoRoleName.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Linktal Recruitment API
 * API for Linktal Recruitment OS
 * OpenAPI spec version: 1.0
 */

export type UpdateConsultantDtoRoleName = typeof UpdateConsultantDtoRoleName[keyof typeof UpdateConsultantDtoRoleName];


export const UpdateConsultantDtoRoleName = {
  admin: 'admin',
  manager: 'manager',
  consultant: 'consultant',
  finance: 'finance',
  researcher: 'researcher',
  viewer: 'viewer',
} as const;